+++ src/webviews/documentdb/collectionView/components/queryInsightsTab/components/index.ts
@@ -4,6 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 export * from './animatedCardList';
+export * from './FeedbackCard';
+export * from './FeedbackDialog';
 export * from './optimizationCards';
 export * from './queryPlanSummary';
 export * from './QuickActions';